model_key and BaseResource#record_path
