+++ app/components/avo/fields/trix_field/edit_component.html.erb
@@ -2,7 +2,7 @@
   <div
     data-controller="trix-field"
     data-trix-field-target="controller"
-    data-resource-name="<%= @resource.model_class.model_name.route_key %>"
+    data-resource-name="<%= @resource.model_key %>"
     data-resource-id="<%= @resource.model.id %>"
     data-attachments-disabled="<%= @field.attachments_disabled %>"
     data-attachment-key="<%= @field.attachment_key %>"
