+++ app/helpers/avo/resources_helper.rb
@@ -42,7 +42,7 @@ module Avo
     end
 
     def item_selector_init(resource)
-      "data-resource-name='#{resource.model_class.model_name.route_key}' data-resource-id='#{resource.model.id}' data-controller='item-selector'"
+      "data-resource-name='#{resource.model_key}' data-resource-id='#{resource.model.id}' data-controller='item-selector'"
     end
 
     def item_selector_input(floating: false, size: :md)
