+++ app/views/avo/actions/show.html.erb
@@ -3,11 +3,11 @@
     data-controller="action"
     data-no-confirmation="<%= @action.no_confirmation %>"
     data-action-target="controllerDiv"
-    data-resource-name="<%= @resource.model_class.model_name.route_key %>"
+    data-resource-name="<%= @resource.model_key %>"
     data-resource-id="<%= params[:id] %>"
     class="hidden text-blue-gray-800"
   >
-    <%= form_with model: @model, scope: 'fields', url: "#{avo.root_path}resources/#{@resource.model_class.model_name.route_key}/actions/#{@action.param_id}", data: {'turbo-frame': '_top', 'action-target': 'form'} do |form| %>
+    <%= form_with model: @model, scope: 'fields', url: "#{@resource.records_path}/actions/#{@action.param_id}", data: {'turbo-frame': '_top', 'action-target': 'form'} do |form| %>
       <%= render Avo::ModalComponent.new do |c| %>
         <% c.heading do %>
           <%= @action.action_name %>
