+++ app/views/avo/base/_actions.html.erb
@@ -14,8 +14,8 @@
       <%
         @actions.each_with_index do |action, index|
           path = action_name == 'show' ?
-            "#{avo.root_path}resources/#{@resource.model_class.model_name.route_key}/#{@model.id}/actions/#{action.param_id}" :
-            "#{avo.root_path}resources/#{@resource.model_class.model_name.route_key}/actions/#{action.param_id}"
+            "#{@resource.record_path}/actions/#{action.param_id}" :
+            "#{@resource.records_path}/actions/#{action.param_id}"
           if action_name == 'show' || action.standalone
             disabled = false
           else
