+++ lib/avo/fields/has_one_field.rb
@@ -1,6 +1,6 @@
 module Avo
   module Fields
-    class HasOneField < BaseField
+    class HasOneField < HasBaseField
       attr_accessor :relation_method
       attr_accessor :display
 
@@ -21,24 +21,6 @@ module Avo
 
       def resource
         Avo::App.get_resource_by_model_name @model.class
-      end
-
-      def turbo_frame
-        "#{self.class.name.demodulize.to_s.underscore}_#{display}_#{id}"
-      end
-
-      def frame_url
-        "#{Avo::App.root_path}/resources/#{resource.model_class.model_name.route_key}/#{@model.id}/#{id}/#{value.id}?turbo_frame=#{turbo_frame}"
-      end
-
-      def target_resource
-        if @model._reflections[id.to_s].klass.present?
-          Avo::App.get_resource_by_model_name @model._reflections[id.to_s].klass.to_s
-        elsif @model._reflections[id.to_s].options[:class_name].present?
-          Avo::App.get_resource_by_model_name @model._reflections[id.to_s].options[:class_name]
-        else
-          Avo::App.get_resource_by_name id.to_s
-        end
       end
 
       def fill_field(model, key, value, params)
